API/pyproject.toml
[project]
name = "API"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-openai>=1.1.11",
    "langserve>=0.3.3",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "fastapi",
    "uvicorn",
    "sse-starlette>=3.3.3",
    "langchain>=1.2.13",
    "langsmith>=0.7.22",
]
